Write a program that prints the first 50 Fibonacci numbers, starting with 1 and 2, followed by a new line.

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,15 +6,16 @@
  **/
 int main(void)
 {
-	int prev1 = 1;  /** First number **/
-	int prev2 = 2;  /** Second number **/
+	long int prev1 = 1;  /** First number **/
+	long int prev2 = 2;  /** Second number **/
 
-	printf("%d, %d", prev1, prev2);
+	printf("%ld, %ld", prev1, prev2);
 
 	for (int i = 2; i < 50; i++)
 	{
-	int current = prev1 + prev2;
-	printf(", %d", current);
+	long int current = prev1 + prev2;
+
+	printf(", %ld", current);
 
 	prev1 = prev2;	/** So that the next num takes the value of the previous num**/
 	prev2 = current;
